Backend/requirements.txt
SQLAlchemy==2.0.21
psycopg2==2.9.7
pandas==2.1.1
python-decouple==3.8